Move intervention action views to separate file
